# api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "KIN API",
    "version": "1.0.0",
    "description": "KIN AI Companion Platform — We Build You A Friend\n\nThe KIN API powers Mission Control (web dashboard), Telegram/Discord/WhatsApp bots, local KIN instances, and third-party integrations. It provides:\n\n- **Auth** — Telegram Login Widget + JWT-based sessions\n- **Chat** — Multi-companion AI chat with a two-brain supervisor (Local Ollama → Groq → Anthropic/OpenAI fallback)\n- **Skills Marketplace** — Browse, install, and custom-submit companion skill extensions\n- **Companion Skills** — NFT-portable XP/leveling system for companion skill trees\n- **Heartbeat & Recovery** — Local KIN health monitoring and snapshot-based recovery\n- **Billing** — Stripe subscription checkout, portal, and webhook handling\n- **Support** — AI-powered support chat with human escalation, FAQ, tickets, and feature requests\n\nAll protected routes require a JWT bearer token obtained via `POST /auth/telegram` or `POST /auth/dev-login` (development only).",
    "contact": {
      "name": "KIN Support",
      "url": "https://www.meetyourkin.com/support",
      "email": "support@meetyourkin.com"
    },
    "license": {
      "name": "MIT",
      "url": "https://github.com/kr8tiv-ai/kr8tiv-runtime-truth-contracts/blob/main/LICENSE"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Development"
    },
    {
      "url": "https://api.meetyourkin.com",
...
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT access token. Obtain via `POST /auth/telegram` or `POST /auth/dev-login`. Token expires after 2 hours. Algorithm: HS256."
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "required": ["error"],
        "properties": {
          "error": {
            "oneOf": [
              { "type": "string" },
              {
                "type": "object",
                "properties": {
                  "message": { "type": "string" },
                  "statusCode": { "type": "integer" }
                }
              }
            ]
          }
        },
        "example": { "error": "Unauthorized" }
      },
      "ValidationError": {
        "type": "object",
        "properties": {
          "error": { "type": "string" },
          "message": { "type": "string" }
        },
        "example": { "error": "Bad Request", "message": "Message is required" }
      },
...
      "CompanionId": {
        "type": "string",
        "enum": ["cipher", "mischief", "vortex", "forge", "aether", "catalyst"],
        "description": "Companion identifier. Each companion has a distinct personality and specialisation.",
        "example": "cipher"
      },
...
      "Skill": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "example": "skill-uuid" },
          "name": { "type": "string", "example": "code_review" },
          "displayName": { "type": "string", "example": "Code Review" },
          "description": { "type": "string", "example": "Reviews code for bugs, style, and performance." },
          "category": { "type": "string", "example": "development" },
          "sourceType": {
            "type": "string",
            "enum": ["built_in", "community", "custom"],
            "example": "built_in"
          },
          "author": { "type": "string", "nullable": true, "example": "KIN Team" },
          "installCount": { "type": "integer", "example": 1024 },
          "version": { "type": "string", "example": "1.0.0" },
          "isInstalled": { "type": "boolean" },
          "isActive": { "type": "boolean" }
        }
      },
...
    },
    "responses": {
      "Unauthorized": {
        "description": "Missing or invalid JWT token.",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/Error" },
            "example": { "error": "Unauthorized" }
          }
        }
      },
      "BadRequest": {
        "description": "Invalid request body or parameters.",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ValidationError" }
          }
        }
      },
      "NotFound": {
        "description": "Resource not found.",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/Error" },
            "example": { "error": "Not Found" }
          }
        }
      },
      "TooManyRequests": {
        "description": "Rate limit exceeded. Default: 100 requests/minute per IP in production.",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "statusCode": { "type": "integer", "example": 429 },
                "error": { "type": "string", "example": "Too Many Requests" },
                "message": { "type": "string" }
              }
            }
          }
        }
      },
      "Conflict": {
        "description": "Conflict — resource already exists.",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/Error" }
          }
        }
      }
    }
  },
  "security": [],
...
  "paths": {
...
    "/skills": {
      "get": {
        "tags": ["Skills"],
        "operationId": "skillsList",
        "summary": "List skill catalog",
        "description": "Returns all approved skills in the marketplace. Each skill includes an `isInstalled` and `isActive` flag relative to the authenticated user. Optionally filter by category or search term.",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "category",
            "in": "query",
            "description": "Filter by skill category (e.g. `development`, `productivity`, `creative`).",
            "schema": { "type": "string", "example": "development" }
          },
          {
            "name": "search",
            "in": "query",
            "description": "Full-text search across skill name and description.",
            "schema": { "type": "string", "example": "code review" }
          }
        ],
        "responses": {
          "200": {
            "description": "Array of skills.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/Skill" }
                }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "429": { "$ref": "#/components/responses/TooManyRequests" }
        }
      }
    },
    "/skills/mine": {
      "get": {
        "tags": ["Skills"],
        "operationId": "skillsMine",
        "summary": "User's installed skills",
        "description": "Returns all skills installed by the authenticated user, including which companion each skill is assigned to and when it was installed.",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "User's installed skills with companion assignment.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "allOf": [
                      { "$ref": "#/components/schemas/Skill" },
                      {
                        "type": "object",
                        "properties": {
                          "companionId": {
                            "$ref": "#/components/schemas/CompanionId",
                            "nullable": true
                          },
                          "installedAt": { "type": "string", "format": "date-time" }
                        }
                      }
                    ]
                  }
                }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "429": { "$ref": "#/components/responses/TooManyRequests" }
        }
      }
    },
    "/skills/{id}/toggle": {
      "post": {
        "tags": ["Skills"],
        "operationId": "skillToggle",
        "summary": "Toggle skill on/off",
        "description": "Installs a skill for the first time (if not yet installed) and sets its active state. If the skill is already installed, only the `is_active` flag is toggled. On first install, the skill's global install count is incremented.",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Skill ID.",
            "schema": { "type": "string", "example": "skill-abc123" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["active"],
                "properties": {
                  "active": {
                    "type": "boolean",
                    "description": "Desired active state for the skill."
                  },
                  "companionId": {
                    "$ref": "#/components/schemas/CompanionId",
                    "description": "Companion to associate the skill with. Optional.",
                    "nullable": true
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Toggle successful.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean", "example": true },
                    "active": { "type": "boolean" }
                  }
                }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "429": { "$ref": "#/components/responses/TooManyRequests" }
        }
      }
    },
    "/skills/request": {
      "post": {
        "tags": ["Skills"],
        "operationId": "skillRequestCreate",
        "summary": "Submit GitHub repo for review",
        "description": "Submit a GitHub repository URL for custom skill review. A $4.99 review fee is charged via Stripe. The repository is reviewed by the KIN team and, if approved, published to the marketplace.\n\nReturns the request ID and `amountCents`. Follow up with `POST /skills/request/{id}/checkout` to pay the review fee.",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["githubRepoUrl"],
                "properties": {
                  "githubRepoUrl": {
                    "type": "string",
                    "format": "uri",
                    "description": "Full GitHub repository URL. Must be a public repository.",
                    "example": "https://github.com/alice/my-kin-skill"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Skill request created. Proceed to checkout.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": { "type": "string", "example": "sr-uuid" },
                    "status": { "type": "string", "enum": ["payment_required"], "example": "payment_required" },
                    "amountCents": { "type": "integer", "example": 499 }
                  }
                }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "409": { "$ref": "#/components/responses/Conflict" },
          "429": { "$ref": "#/components/responses/TooManyRequests" }
        }
      }
    },
    "/skills/request/{id}/checkout": {
      "post": {
        "tags": ["Skills"],
        "operationId": "skillRequestCheckout",
        "summary": "Stripe checkout for $4.99 review fee",
        "description": "Creates a Stripe Checkout session for the one-time $4.99 skill review fee. Returns the Stripe Checkout URL to redirect the user to. On successful payment, the webhook updates the skill request status to `paid` and triggers the review queue.",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Skill request ID returned from `POST /skills/request`.",
            "schema": { "type": "string", "example": "sr-uuid" }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "successUrl": {
                    "type": "string",
                    "format": "uri",
                    "description": "URL to redirect to after payment. Defaults to the KIN dashboard.",
                    "example": "https://www.meetyourkin.com/dashboard/skills?request=submitted"
                  },
                  "cancelUrl": {
                    "type": "string",
                    "format": "uri",
                    "description": "URL to redirect to on payment cancellation.",
                    "example": "https://www.meetyourkin.com/dashboard/skills"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Stripe Checkout URL.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "url": {
                      "type": "string",
                      "format": "uri",
                      "nullable": true,
                      "description": "Redirect the user to this URL. Null if Stripe is not configured."
                    },
                    "message": { "type": "string", "description": "Present if Stripe is not configured." }
                  }
                }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "429": { "$ref": "#/components/responses/TooManyRequests" }
        }
      }
    },
    "/skills/request/mine": {
      "get": {
        "tags": ["Skills"],
        "operationId": "skillRequestsMine",
        "summary": "User's skill requests",
        "description": "Returns all skill requests submitted by the authenticated user, ordered by submission date descending.",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "List of skill requests.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": { "type": "string" },
                      "githubRepoUrl": { "type": "string", "format": "uri" },
                      "repoOwner": { "type": "string" },
                      "repoName": { "type": "string" },
                      "skillName": { "type": "string", "nullable": true },
                      "skillDescription": { "type": "string", "nullable": true },
                      "status": {
                        "type": "string",
                        "enum": ["payment_required", "paid", "under_review", "approved", "rejected", "installed"]
                      },
                      "rejectionReason": { "type": "string", "nullable": true },
                      "amountCents": { "type": "integer" },
                      "createdAt": { "type": "string", "format": "date-time" },
                      "updatedAt": { "type": "string", "format": "date-time" }
                    }
                  }
                }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "429": { "$ref": "#/components/responses/TooManyRequests" }
        }
      }
    },
...
  }
}